LayoutTests/svg/dynamic-updates/SVGForeignObjectElement-dom-requiredFeatures.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="resources/SVGTestCase.js"></script>
<script src="../../resources/js-test-pre.js"></script>
<script src="../../fast/repaint/resources/repaint.js"></script>
</head>
<body onload="runRepaintTest()">
<h1>SVG 1.1 dynamic update tests</h1>
<p id="description"></p>
<div id="console"></div>
<script>
// [Name] SVGForeignObjectElement-dom-requiredFeatures.js
// [Expected rendering result] a series of PASS messages

createSVGTestCase();

var foreignObjectElement = createSVGElement("foreignObject");
foreignObjectElement.setAttribute("width", "200");
foreignObjectElement.setAttribute("height", "200");

var htmlDivElement = document.createElementNS(xhtmlNS, "xhtml:div");
htmlDivElement.setAttribute("style", "background-color: green; color: white; text-align: center");
htmlDivElement.textContent = "Test passed";

foreignObjectElement.appendChild(htmlDivElement);
rootSVGElement.appendChild(foreignObjectElement);

function repaintTest() {
    debug("Check that SVGForeignObjectElement is initially displayed");
    shouldBeEqualToString("document.defaultView.getComputedStyle(foreignObjectElement, null).display", "block");
    debug("Check that setting requiredFeatures to something invalid makes it not render");
    foreignObjectElement.setAttribute("requiredFeatures", "http://www.w3.org/TR/SVG11/feature#BogusFeature");
    shouldBeEqualToString("document.defaultView.getComputedStyle(foreignObjectElement, null).display", "");
    debug("Check that setting requiredFeatures to something valid makes it render again");
    foreignObjectElement.setAttribute("requiredFeatures", "http://www.w3.org/TR/SVG11/feature#Shape");
    shouldBeEqualToString("document.defaultView.getComputedStyle(foreignObjectElement, null).display", "block");
    debug("Check that adding something valid to requiredFeatures keeps rendering the element");
    foreignObjectElement.setAttribute("requiredFeatures", "http://www.w3.org/TR/SVG11/feature#Gradient");
    shouldBeEqualToString("document.defaultView.getComputedStyle(foreignObjectElement, null).display", "block");
    debug("Check that adding something invalid to requiredFeatures makes it not render");
    foreignObjectElement.setAttribute("requiredFeatures", "http://www.w3.org/TR/SVG11/feature#BogusFeature");
    shouldBeEqualToString("document.defaultView.getComputedStyle(foreignObjectElement, null).display", "");

    completeTest();
}

var successfullyParsed = true;
</script>
</body>
</html>
